requirements.txt
yfinance==0.2.43
pandas==2.2.2
numpy==1.26.4
torch==2.4.1
transformers==4.44.2
requests==2.32.3
beautifulsoup4==4.12.3
newspaper3k==0.2.8
rich==13.8.0
sqlite3  # Built-in, but list for clarity